requirements.txt
python-dotenv
mysqlclient
pymysql
djangorestframework
djangorestframework-simplejwt
Pillow
graphene
graphene_django